+++ Cargo.toml
@@ -76,7 +76,6 @@ hmac = "0.12"
 sha2 = "0.10"
 bitvec = "1"
 flate2 = "1"
-enum_dispatch = "0.3"
 petgraph = "0.8"
 thiserror = "2"
 syn = "2"
@@ -96,3 +95,4 @@ anstream = "1.0"
 owo-colors = "4.3"
 smallvec = { version = "1.15", features = ["union", "serde"] }
 rayon = "1.12"
+enum_delegate = "0.2.0"
