+++ crates/backends/direct/src/tick.rs
@@ -1,3 +1,5 @@
+use mchprs_backend_lib::bool_to_ss;
+
 use super::node::NodeId;
 use super::*;
